+++ Example/Sources/Settings+UserDefaults.swift
@@ -1,7 +1,7 @@
 /*
  MIT License
  
- Copyright (c) 2017-2018 MessageKit
+ Copyright (c) 2017 MessageKit
  
  Permission is hereby granted, free of charge, to any person obtaining a copy
  of this software and associated documentation files (the "Software"), to deal
@@ -24,10 +24,21 @@
 
 import Foundation
 
-extension CGRect {
+extension UserDefaults {
     
-    init(_ x: CGFloat, _ y: CGFloat, _ w: CGFloat, _ h: CGFloat) {
-        self.init(x: x, y: y, width: w, height: h)
+    static let messagesKey = "mockMessages"
+    
+    // MARK: - Mock Messages
+    
+    func setMockMessages(count: Int) {
+        set(count, forKey: "mockMessages")
+        synchronize()
     }
-
+    
+    func mockMessagesCount() -> Int {
+        if let value = object(forKey: "mockMessages") as? Int {
+            return value
+        }
+        return 20
+    }
 }