+++ Sources/Controllers/MessagesViewController.swift
@@ -1,4 +1,3 @@
-
 /*
  MIT License
 
@@ -58,11 +57,14 @@ open class MessagesViewController: UIViewController {
     open override var shouldAutorotate: Bool {
         return false
     }
-    
+
     /// A Boolean value used to determine if `viewDidLayoutSubviews()` has been called.
     private var isFirstLayout: Bool = true
     
-    private var messageCollectionViewBottomInset: CGFloat = 0 {
+    /// Indicated selected indexPath when handle menu action
+    var selectedIndexPathForMenu: IndexPath?
+
+    var messageCollectionViewBottomInset: CGFloat = 0 {
         didSet {
             messagesCollectionView.contentInset.bottom = messageCollectionViewBottomInset
             messagesCollectionView.scrollIndicatorInsets.bottom = messageCollectionViewBottomInset
@@ -73,18 +75,12 @@ open class MessagesViewController: UIViewController {
 
     open override func viewDidLoad() {
         super.viewDidLoad()
-
-        extendedLayoutIncludesOpaqueBars = true
-        automaticallyAdjustsScrollViewInsets = false
-        view.backgroundColor = .white
-        messagesCollectionView.keyboardDismissMode = .interactive
-        messagesCollectionView.alwaysBounceVertical = true
-        
+        setupDefaults()
         setupSubviews()
         setupConstraints()
         registerReusableViews()
         setupDelegates()
-
+        addMenuControllerObservers()
     }
 
     open override func viewDidLayoutSubviews() {
@@ -101,9 +97,19 @@ open class MessagesViewController: UIViewController {
 
     deinit {
         removeKeyboardObservers()
+        removeMenuControllerObservers()
     }
 
     // MARK: - Methods [Private]
+
+    /// Sets the default values for the MessagesViewController
+    private func setupDefaults() {
+        extendedLayoutIncludesOpaqueBars = true
+        automaticallyAdjustsScrollViewInsets = false
+        view.backgroundColor = .white
+        messagesCollectionView.keyboardDismissMode = .interactive
+        messagesCollectionView.alwaysBounceVertical = true
+    }
 
     /// Sets the delegate and dataSource of the messagesCollectionView property.
     private func setupDelegates() {
@@ -111,9 +117,13 @@ open class MessagesViewController: UIViewController {
         messagesCollectionView.dataSource = self
     }
 
+    /// Adds the messagesCollectionView to the controllers root view.
+    private func setupSubviews() {
+        view.addSubview(messagesCollectionView)
+    }
+
     /// Registers all cells and supplementary views of the messagesCollectionView property.
     private func registerReusableViews() {
-
         messagesCollectionView.register(TextMessageCell.self)
         messagesCollectionView.register(MediaMessageCell.self)
         messagesCollectionView.register(LocationMessageCell.self)
@@ -121,12 +131,6 @@ open class MessagesViewController: UIViewController {
         messagesCollectionView.register(MessageFooterView.self, forSupplementaryViewOfKind: UICollectionElementKindSectionFooter)
         messagesCollectionView.register(MessageHeaderView.self, forSupplementaryViewOfKind: UICollectionElementKindSectionHeader)
         messagesCollectionView.register(MessageDateHeaderView.self, forSupplementaryViewOfKind: UICollectionElementKindSectionHeader)
-
-    }
-
-    /// Adds the messagesCollectionView to the controllers root view.
-    private func setupSubviews() {
-        view.addSubview(messagesCollectionView)
     }
 
     /// Sets the constraints of the `MessagesCollectionView`.
@@ -144,188 +148,5 @@ open class MessagesViewController: UIViewController {
             let trailing = messagesCollectionView.trailingAnchor.constraint(equalTo: view.trailingAnchor)
             NSLayoutConstraint.activate([top, bottom, trailing, leading])
         }
-    }
-    
-    @objc
-    private func adjustScrollViewInset() {
-        if #available(iOS 11.0, *) {
-            // No need to add to the top contentInset
-        } else {
-            var navigationBarInset: CGFloat = 0
-            if let navigationController = navigationController, !navigationController.isNavigationBarHidden {
-                navigationBarInset = navigationController.navigationBar.frame.height
-            }
-            let statusBarInset: CGFloat = UIApplication.shared.isStatusBarHidden ? 0 : 20
-            let topInset = navigationBarInset + statusBarInset
-            messagesCollectionView.contentInset.top = topInset
-            messagesCollectionView.scrollIndicatorInsets.top = topInset
-        }
-    }
-}
-
-// MARK: - UICollectionViewDelegate & UICollectionViewDelegateFlowLayout Conformance
-
-extension MessagesViewController: UICollectionViewDelegateFlowLayout {
-
-    open func collectionView(_ collectionView: UICollectionView, layout collectionViewLayout: UICollectionViewLayout, sizeForItemAt indexPath: IndexPath) -> CGSize {
-        guard let messagesFlowLayout = collectionViewLayout as? MessagesCollectionViewFlowLayout else { return .zero }
-        return messagesFlowLayout.sizeForItem(at: indexPath)
-    }
-    
-    open func collectionView(_ collectionView: UICollectionView, layout collectionViewLayout: UICollectionViewLayout, referenceSizeForHeaderInSection section: Int) -> CGSize {
-        guard let messagesCollectionView = collectionView as? MessagesCollectionView else { return .zero }
-        guard let messagesDataSource = messagesCollectionView.messagesDataSource else { return .zero }
-        guard let messagesLayoutDelegate = messagesCollectionView.messagesLayoutDelegate else { return .zero }
-        // Could pose a problem if subclass behaviors allows more than one item per section
-        let indexPath = IndexPath(item: 0, section: section)
-        let message = messagesDataSource.messageForItem(at: indexPath, in: messagesCollectionView)
-        return messagesLayoutDelegate.headerViewSize(for: message, at: indexPath, in: messagesCollectionView)
-    }
-    
-    open func collectionView(_ collectionView: UICollectionView, layout collectionViewLayout: UICollectionViewLayout, referenceSizeForFooterInSection section: Int) -> CGSize {
-        guard let messagesCollectionView = collectionView as? MessagesCollectionView else { return .zero }
-        guard let messagesDataSource = messagesCollectionView.messagesDataSource else { return .zero }
-        guard let messagesLayoutDelegate = messagesCollectionView.messagesLayoutDelegate else { return .zero }
-        // Could pose a problem if subclass behaviors allows more than one item per section
-        let indexPath = IndexPath(item: 0, section: section)
-        let message = messagesDataSource.messageForItem(at: indexPath, in: messagesCollectionView)
-        return messagesLayoutDelegate.footerViewSize(for: message, at: indexPath, in: messagesCollectionView)
-    }
-
-}
-
-// MARK: - UICollectionViewDataSource Conformance
-
-extension MessagesViewController: UICollectionViewDataSource {
-
-    open func numberOfSections(in collectionView: UICollectionView) -> Int {
-        guard let collectionView = collectionView as? MessagesCollectionView else { return 0 }
-
-        // Each message is its own section
-        return collectionView.messagesDataSource?.numberOfMessages(in: collectionView) ?? 0
-    }
-
-    open func collectionView(_ collectionView: UICollectionView, numberOfItemsInSection section: Int) -> Int {
-        guard let collectionView = collectionView as? MessagesCollectionView else { return 0 }
-
-        let messageCount = collectionView.messagesDataSource?.numberOfMessages(in: collectionView) ?? 0
-        // There will only ever be 1 message per section
-        return messageCount > 0 ? 1 : 0
-
-    }
-
-    open func collectionView(_ collectionView: UICollectionView, cellForItemAt indexPath: IndexPath) -> UICollectionViewCell {
-
-        guard let messagesCollectionView = collectionView as? MessagesCollectionView else {
-            fatalError("Managed collectionView: \(collectionView.debugDescription) is not a MessagesCollectionView.")
-        }
-
-        guard let messagesDataSource = messagesCollectionView.messagesDataSource else {
-            fatalError("MessagesDataSource has not been set.")
-        }
-
-        let message = messagesDataSource.messageForItem(at: indexPath, in: messagesCollectionView)
-
-        switch message.data {
-        case .text, .attributedText, .emoji:
-            let cell = messagesCollectionView.dequeueReusableCell(TextMessageCell.self, for: indexPath)
-            cell.configure(with: message, at: indexPath, and: messagesCollectionView)
-            return cell
-        case .photo, .video:
-    	    let cell = messagesCollectionView.dequeueReusableCell(MediaMessageCell.self, for: indexPath)
-            cell.configure(with: message, at: indexPath, and: messagesCollectionView)
-            return cell
-        case .location:
-    	    let cell = messagesCollectionView.dequeueReusableCell(LocationMessageCell.self, for: indexPath)
-            cell.configure(with: message, at: indexPath, and: messagesCollectionView)
-            return cell
-        }
-    }
-
-    open func collectionView(_ collectionView: UICollectionView, viewForSupplementaryElementOfKind kind: String, at indexPath: IndexPath) -> UICollectionReusableView {
-
-        guard let messagesCollectionView = collectionView as? MessagesCollectionView else {
-            fatalError("Managed collectionView: \(collectionView.debugDescription) is not a MessagesCollectionView.")
-        }
-
-        guard let dataSource = messagesCollectionView.messagesDataSource else {
-            fatalError("MessagesDataSource has not been set.")
-        }
-
-        guard let displayDelegate = messagesCollectionView.messagesDisplayDelegate else {
-            fatalError("MessagesDisplayDelegate has not been set.")
-        }
-
-        let message = dataSource.messageForItem(at: indexPath, in: messagesCollectionView)
-
-        switch kind {
-        case UICollectionElementKindSectionHeader:
-            return displayDelegate.messageHeaderView(for: message, at: indexPath, in: messagesCollectionView)
-        case UICollectionElementKindSectionFooter:
-            return displayDelegate.messageFooterView(for: message, at: indexPath, in: messagesCollectionView)
-        default:
-            fatalError("Unrecognized element of kind: \(kind)")
-        }
-    }
-}
-
-// MARK: - Keyboard Handling
-
-fileprivate extension MessagesViewController {
-
-    func addKeyboardObservers() {
-        NotificationCenter.default.addObserver(self, selector: #selector(handleKeyboardDidChangeState), name: .UIKeyboardWillChangeFrame, object: nil)
-        NotificationCenter.default.addObserver(self, selector: #selector(handleTextViewDidBeginEditing), name: .UITextViewTextDidBeginEditing, object: nil)
-    }
-
-    func removeKeyboardObservers() {
-        NotificationCenter.default.removeObserver(self, name: .UIKeyboardWillChangeFrame, object: nil)
-        NotificationCenter.default.removeObserver(self, name: .UITextViewTextDidBeginEditing, object: nil)
-    }
-
-    @objc
-    func handleTextViewDidBeginEditing(_ notification: Notification) {
-        if scrollsToBottomOnKeybordBeginsEditing {
-            guard let inputTextView = notification.object as? InputTextView, inputTextView === messageInputBar.inputTextView else { return }
-            messagesCollectionView.scrollToBottom(animated: true)
-        }
-    }
-
-    @objc
-    func handleKeyboardDidChangeState(_ notification: Notification) {
-        guard let keyboardEndFrame = notification.userInfo?[UIKeyboardFrameEndUserInfoKey] as? CGRect else { return }
-
-        if (keyboardEndFrame.origin.y + keyboardEndFrame.size.height) > UIScreen.main.bounds.height {
-            // Hardware keyboard is found
-            messageCollectionViewBottomInset = view.frame.size.height - keyboardEndFrame.origin.y - iPhoneXBottomInset
-        } else {
-            //Software keyboard is found
-            let afterBottomInset = keyboardEndFrame.height > keyboardOffsetFrame.height ? (keyboardEndFrame.height - iPhoneXBottomInset) : keyboardOffsetFrame.height
-            let differenceOfBottomInset = afterBottomInset - messageCollectionViewBottomInset
-            let contentOffset = CGPoint(x: messagesCollectionView.contentOffset.x, y: messagesCollectionView.contentOffset.y + differenceOfBottomInset)
-
-            if maintainPositionOnKeyboardFrameChanged {
-                messagesCollectionView.setContentOffset(contentOffset, animated: false)
-            }
-
-            messageCollectionViewBottomInset = afterBottomInset
-        }
-    }
-    
-    fileprivate var keyboardOffsetFrame: CGRect {
-        guard let inputFrame = inputAccessoryView?.frame else { return .zero }
-        return CGRect(origin: inputFrame.origin, size: CGSize(width: inputFrame.width, height: inputFrame.height - iPhoneXBottomInset))
-    }
-    
-    /// On the iPhone X the inputAccessoryView is anchored to the layoutMarginesGuide.bottom anchor so the frame of the inputAccessoryView
-    /// is larger than the required offset for the MessagesCollectionView
-    ///
-    /// - Returns: The safeAreaInsets.bottom if its an iPhoneX, else 0
-    fileprivate var iPhoneXBottomInset: CGFloat {
-        if #available(iOS 11.0, *) {
-            guard UIScreen.main.nativeBounds.height == 2436 else { return 0 }
-            return view.safeAreaInsets.bottom
-        }
-        return 0
     }
 }