+++ Tests/ProtocolsTests/MessagesDisplayDelegateTests.swift
@@ -85,12 +85,6 @@ class MessagesDisplayDelegateTests: XCTestCase {
         XCTAssertEqual(backgroundColor, .clear)
     }
 
-    func testAvatarDefaultState() {
-        XCTAssertNotNil(sut.dataProvider.avatar(for: sut.dataProvider.messages[0],
-                                                at: IndexPath(item: 0, section: 0),
-                                                in: sut.messagesCollectionView).initials)
-    }
-
     func testCellTopLabelDefaultState() {
         XCTAssertNil(sut.dataProvider.cellTopLabelAttributedText(for: sut.dataProvider.messages[0],
                                                                  at: IndexPath(item: 0, section: 0)))
@@ -218,7 +212,8 @@ class TextMessageDisplayDelegateTests: XCTestCase {
     }
 
     func testTextColorWithoutDataSource_returnsDarkTextForDefault() {
-        sut.messagesCollectionView.messagesDataSource = nil
+        let dataSource = sut.makeDataSource()
+        sut.messagesCollectionView.messagesDataSource = dataSource
         let textColor = sut.textColor(for: sut.dataProvider.messages[1],
                                       at: IndexPath(item: 0, section: 0),
                                       in: sut.messagesCollectionView)
@@ -255,7 +250,7 @@ private class MockMessagesViewController: MessagesViewController, MessagesDispla
 
     }
 
-    private func makeDataSource() -> MockMessagesDataSource {
+    fileprivate func makeDataSource() -> MockMessagesDataSource {
         let dataSource = MockMessagesDataSource()
         dataSource.messages.append(MockMessage(text: "Text 1",
                                                sender: dataSource.senders[0],
@@ -267,4 +262,7 @@ private class MockMessagesViewController: MessagesViewController, MessagesDispla
         return dataSource
     }
 
+    func snapshotOptionsForLocation(message: MessageType, at indexPath: IndexPath, in messagesCollectionView: MessagesCollectionView) -> LocationMessageSnapshotOptions {
+        return LocationMessageSnapshotOptions()
+    }
 }